CRUD for batch

+++ batch/views/delete_batch.py
@@ -1,11 +1,16 @@
+from django.http import Http404
 from django.shortcuts import redirect
 from batch.models import Batch
 from django.shortcuts import redirect, render
+from django.http import Http404
+
 
 def deleteBatch(request,id):
-    context = {}
     batch = Batch.get_batch_by_id(id)
-    if request.method == "POST":
-        batch.delete()
-        return redirect('batch:list_batch')
-    return render(request,'batchDelete.html')
+    if batch:
+        if request.method == "POST":
+            batch.delete()
+            return redirect('batch:list_batch')
+        return render(request,'batch/batchDelete.html')
+    else:
+        raise Http404()
